DataLoader puesto + correccion en el modelo y repositorio

+++ src/main/java/com/example/netpick_back/demo/service/CategoriaService.java
@@ -49,6 +49,6 @@ public class CategoriaService {
     }
 
     public void deleteById(Integer id) {
-        categoriaRepository.deleteById(id.longValue());
+        categoriaRepository.deleteById(id.intValue());
     }
 }
